+++ processor/src/main/java/arez/processor/TrackedDescriptor.java
@@ -158,7 +158,7 @@ final class TrackedDescriptor
     parameters.add( _componentDescriptor.getComponentNameMethodName() );
     parameters.add( "." + getName() );
     sb.append( _mutation );
-    sb.append( ", () -> super.$N() )" );
+    sb.append( ", () -> super.$N(), false )" );
     parameters.add( _onDepsChangedMethod.getSimpleName().toString() );
 
     builder.addStatement( sb.toString(), parameters.toArray() );